Fixed issues with Neo Forge and Forge version and generated json (#134)
Bumped Kotlin for forge version which fixes issue with trident not being able to be thrown.
Fixed issues with some item and armour materials not being registered properly.
Updated json files to reflect Neo Forges changes.
Identified issue with trident not being able to be thrown. Will be fixed in future.
Fixed issues with some item and armour materials not being registered properly.